+++ src/extension.ts
@@ -6,13 +6,13 @@ import { PyDataCustomProvider } from './pydataProvider';
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-	
+
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
 	console.log('Congratulations, your extension "vscode-pydata-viewer" is now active!');
 
 	const extensionRoot = vscode.Uri.file(context.extensionPath);
-	const provider = new PyDataCustomProvider(extensionRoot);
+	const provider = new PyDataCustomProvider(context, extensionRoot);
 	context.subscriptions.push(
 		vscode.window.registerCustomEditorProvider(
 			PyDataCustomProvider.viewType,
@@ -27,4 +27,4 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }